+++ CHANGES.md
@@ -68,6 +68,8 @@
 
 ## Fixed
 
+- Gospel preprocessor support documentation for ghost declaration
+  [\#331](https://github/ocaml-gospel/gospel/pull/331)
 - Consider comments as spaces while preprocessing (to ensure specification can
   be attached to a ghost function or type, for instance)
   [\#321](https://github/ocaml-gospel/gospel/pull/321)